annotation and add extra blank line

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,8 @@ import random
 import asyncio
 from typing import Generator
 
-async def async_generator() -> Generator[int, None, None]:
+
+async def async_generator() -> Generator[float, None, None]:
     """ Doc """
     for i in range(10):
         await asyncio.sleep(1)
